requirements.txt
google-cloud-storage>=0.34.0
wand
